# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "用户管理系统",
        "contact": {},
        "version": "0.1"
    },
    "paths": {
...
        "/api/submit/query": {
            "get": {
                "description": "get question submit result",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "QuestionSubmit"
                ],
                "summary": "get question submit result",
                "parameters": [
                    {
                        "type": "string",
                        "description": "query",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "query  successful",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/modelquestion.QSReturn"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "query failed",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "get question submit result",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "QuestionSubmit"
                ],
                "summary": "get question submit result",
                "parameters": [
                    {
                        "description": "queries",
                        "name": "query",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modelquestion.QueryQuestionSubmitRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "query  successful",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/modelquestion.QSReturn"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "query failed",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/submit/submission": {
            "post": {
                "description": "submit",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "QuestionSubmit"
                ],
                "summary": "submit",
                "parameters": [
                    {
                        "description": "submit code",
                        "name": "modelquestion",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modelquestion.AddQuestionSubmitRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "submit success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "submit failed",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "modelquestion.AddQuestionSubmitRequest": {
            "type": "object",
            "required": [
                "question_id"
            ],
            "properties": {
                "code": {
                    "description": "\"用户代码\"",
                    "type": "string"
                },
                "language": {
                    "description": "\"编程语言\"",
                    "type": "string"
                },
                "question_id": {
                    "description": "\"题目id\"",
                    "type": "string"
                }
            }
        },
...
        "modelquestion.JudgeInfo": {
            "type": "object",
            "properties": {
                "memory": {
                    "description": "单位为kb",
                    "type": "integer"
                },
                "message": {
                    "description": "值为以上枚举值",
                    "type": "string"
                },
                "time": {
                    "description": "单位为ms",
                    "type": "integer"
                }
            }
        },
        "modelquestion.QSReturn": {
            "type": "object",
            "properties": {
                "judge_info": {
                    "description": "\"判题信息json对象(包含上面的枚举值)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/modelquestion.JudgeInfo"
                        }
                    ]
                },
                "language": {
                    "description": "\"编程语言\"",
                    "type": "string"
                },
                "question_id": {
                    "description": "题目id",
                    "type": "string"
                },
                "status": {
                    "description": "\"判题状态（0-待判题,1-判题中,2-成功,3-失败)\",",
                    "type": "integer"
                }
            }
        },
...
        "modelquestion.QueryQuestionSubmitRequest": {
            "type": "object",
            "properties": {
                "language": {
                    "description": "\"编程语言\"",
                    "type": "string"
                },
                "question_id": {
                    "description": "\"判题id\"",
                    "type": "string"
                },
                "status": {
                    "description": "\"判题状态（0-待判题,1-判题中,2-成功,3-失败)\",",
                    "type": "integer"
                },
                "user_id": {
                    "description": "\"创建用户id\"",
                    "type": "string"
                }
            }
        },
...
        "utils.ApiResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}